+++ bin/mcp_server/utils/bash_command_parser.dart
