+++ bin/mcp_server/utils/systemctl_command_parser.dart
